rtl: Change the conditions for releasing the SPC BUSY state
The SPI protocol engine determines that the transfer is complete when
SPC Busy is negated. If the Busy signal is negated at the same time as
the transition to the IDLE state, the received data may not have been
stored in the register yet. So, change the negate condition for BUSY.

Signed-off-by: Daisuke Sasaki <daisuke@spacecubics.com>