rtl: Change the update timing of the receive frame counter
The receive frame counter "fc_rx" operates by tracking the frame
counter "fc".

When a frame has not started, fc_rx does not need to count. However,
it needs to track the initialization of fc, so it needed to track even
when the frame was invalid. his bug caused fc_rx to not be
initialized, so it has been fixed.

Signed-off-by: Daisuke Sasaki <daisuke@spacecubics.com>