rtl: Add swapping logic for sending data to sc-spi-engine
The byte order circuit has been modified to simplify the control
circuit.

The previous specification controlled the data pointer. The new
specification swaps the data itself. This ensures that the internal
control is the same regardless of the byte order mode.

This fix is ​​only applied to the transmit circuit. So, This fix may
cause the receive circuit to not work properly. It must be applied
together with the following commit.

Signed-off-by: Daisuke Sasaki <daisuke@spacecubics.com>