rtl: Change the negate timing of the chip select signal
The SPI chip select signal is negated only at the end of a
transfer. However, the implementation was such that the negation
decision was made at any time when the state machine was in the idle
state.  However, the implementation was such that the negation check
was made at any time when the state machine was in the idle state.

In this implementation, even if a transfer with CSEXTEND=1 is
performed, the Chip Select signal may be negated during the IDLE state
if the next transfer is with CSEXTEND=0. So, the Chip Select signal is
now negated only when returning to the IDLE state.

Signed-off-by: Daisuke Sasaki <daisuke@spacecubics.com>